Progress on #42
This feature is now working on normal vim. When there is an error
typst.vim will populate the QuickFix list and open it. When the
errors have been fixed, the QuickFix list will be closed.

Some problems remain for neovim. I think it is only minor fixes due
to API differences.